Revert "new uupdates"
