--- app/[locale]/articles/page.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const page = () => {
-  return (
-    <div>Articles Page</div>
-  )
-}
-
-export default page
