+++ app/_components/Header.jsx
@@ -2,16 +2,18 @@
 import Image from "next/image";
 import Link from "next/link";
 import { useTranslations } from "next-intl";
-import { useState } from "react";
+import {  useState } from "react";
+import { useLocale } from "../layout";
 import { usePathname } from "next/navigation";
-// import { useRouter } from "next/navigation";
-// import logo from "@/public/logo3.png";
+
+import logo from "@/public/logo3.png";
 import { GrLanguage } from "react-icons/gr";
 
 export default function Header() {
+  const { currentLocale, setLocale } = useLocale();
 
   const [isMenuOpen, setIsMenuOpen] = useState(false);
-  const [openDropdown, setOpenDropdown] = useState(null);
+  const [openDropdown, setOpenDropdown] = useState<string | null>(null);
 
   const currentPath = usePathname();
 
@@ -31,11 +33,11 @@ export default function Header() {
   const toggleMenu = () => {
     setIsMenuOpen((prev) => !prev);
   };
-  const toggleDropdown = (dropdown) => {
+  const toggleDropdown = (dropdown: string) => {
     setOpenDropdown(openDropdown === dropdown ? null : dropdown);
   };
 
-  const updateLanguage = (newLanguage) => {
+  const updateLanguage = (newLanguage: string) => {
     setLocale(newLanguage);
   };
 
@@ -52,7 +54,7 @@ export default function Header() {
               onClick={() => updateLanguage(currentLocale === "ar" ? "en" : "ar")}
               className="text-gray-600 hover:underline font-bold px-6"
             >
-              {/* {t(locale === "ar" ? "english" : "arabic")} */}
+              {t(currentLocale === "ar" ? "english" : "arabic")}
             </button>
 
             <button className="font-bold h-full bg-primary-green text-white w-24 py-3  rounded-full">
@@ -61,7 +63,7 @@ export default function Header() {
           </div>
 
 
-          <div className="block lg:hidden ps-6 text-black" onClick={toggleMenu}>
+          <div className="block lg:hidden ps-6" onClick={toggleMenu}>
             <svg
               className="w-6 h-6 cursor-pointer"
               xmlns="http://www.w3.org/2000/svg"
@@ -132,12 +134,9 @@ export default function Header() {
           <div className="logo ">
             <Link href="/" className="">
               <Image
-                src="/logo3.png"
+                src={logo}
                 alt="MyCash Logo"
                 className={`object-contain  xxs:h-8 xs:h-10 sm:h-12 md:h-12 xxs:w-40 `}
-                width={200}
-                height={200}
-
               />
             </Link>
           </div>
