+++ app/_components/Hero.jsx
@@ -1,4 +1,3 @@
-
 import React from 'react'
 import Header from './Header'
 import { useTranslations } from 'next-intl'
@@ -6,7 +5,7 @@ import { useTranslations } from 'next-intl'
 const Hero = () => {
     const t = useTranslations("mainpage");
     return (
-        <div className="bg-hero bg-cover bg-no-repeat bg-center w-full aspect-square md:h-[70vh]">
+        <div className="bg-herobg bg-cover bg-no-repeat bg-center w-full aspect-square md:h-[70vh]">
 
             <Header />
             <div className="flex flex-col justify-center items-center gap-6 w-full h-full text-center">
