+++ app/_components/Solutions.jsx
@@ -4,11 +4,15 @@ import React from "react";
 import Link from "next/link";
 import { useTranslations } from "use-intl";
 import { TripleHeadings } from "./TripleHeadings";
+import ai_solution from "@/public/ai_solution.png";
+import pos_solutions from "@/public/pos_solutions.png";
+import sales_solutions from "@/public/sales_solutions.png";
 
-// import ai_solution from "@/public/ai_solution.png";
-// import pos_solutions from "@/public/pos_solutions.png";
-// import sales_solutions from "@/public/sales_solutions.png";
-
+// Images test 
+// import a1 from "../../public/Curved/pos/a1";
+// import a2 from "../../public/Curved/pos/a2"
+// import a3 from "../../public/Curved/pos/a3"
+// import a4 from "../../public/Curved/pos/a4"
 const Solutions = () => {
   const t = useTranslations("solutions");
   return (
@@ -29,10 +33,8 @@ const Solutions = () => {
 
           <div className="rounded-2xl m-auto pt-4 px-4">
             <Image
-              src="/ai_solution.png"
+              src={ai_solution}
               alt="Solutions image"
-              width={500} height={500}
-
               className=" w-full h-full rounded-2xl"
             />
           </div>
@@ -70,9 +72,8 @@ const Solutions = () => {
 
           <div className="rounded-2xl m-auto pt-4 px-4">
             <Image
-              src="/sales_solutions.png"
+              src={sales_solutions}
               alt="Solutions image"
-              width={500} height={500}
               className=" w-full h-full rounded-2xl"
             />
           </div>
@@ -111,10 +112,8 @@ const Solutions = () => {
 
           <div className="rounded-2xl m-auto pt-4 px-4 ">
             <Image
-              src="/pos_solutions.png"
+              src={pos_solutions}
               alt="Solutions image"
-              width={500} height={500}
-
               className=" w-full  rounded-2xl"
             />
           </div>
