+++ app/articles/page.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const page = () => {
+  return (
+    <div>Articles Page</div>
+  )
+}
+
+export default page
