+++ app/globals.css
@@ -2,7 +2,7 @@
 @tailwind components;
 @tailwind utilities;
 
-/* :root {
+:root {
   --background: #ffffff;
   --foreground: #171717;
 }
@@ -19,5 +19,5 @@ body {
   background: var(--background);
   font-family: Arial, Helvetica, sans-serif;
 }
- */
+
