+++ tailwind.config.mjs
@@ -1,4 +1,3 @@
-/* eslint-disable import/no-anonymous-default-export */
 /** @type {import('tailwindcss').Config} */
 export default {
   content: [
@@ -24,7 +23,8 @@ export default {
         backgroundImage:{
           "bg_lg": "url('/cashaicamswide.png')",
           "bg_sm": "url('/mycashaismall.png')",
-          mainbg: "url('/heronew.png')",
+          "mycash_large": "url('/cashaicamswide.png')",
+          "herobg": "url('/heronew.png')",
           "solutions_bg": "url('/solutions_bg.png')",
         },
         screens : {
